Cleanup custom scripts
